Add support for minimum and maximum file counts.
A minimum file count can be used to prevent the launcher from executing the command if a minimal number of files have not been selected.
A maximum file count can be used to execute the command as soon as the maximum number of files have been selected. This is particularly helpful when selecting files individually (e.g. selecting files from separate Windows Explorer instances).